freebsd.mkimg: Support bsd disklabel relocation
Booting OpenBSD on UEFI systems requires a GUID Partition Table
with an EFI System Partition with FAT contents and an "OpenBSD data"
with its own BSD disklabel partition table.
Unfortunately, the offsets in the inner disklabel partition are from
the beginning of the disk, not of the UEFI partition.

Therefore, add support for setting an offset when creating a BSD
disklabel partition

Also fix the name of the "openbsd-ufs" partition type.
It should be "openbsd-data" since it doesn't directly contain
a UFS partition.
The name was incorrect in 83f0e7c88157886d4c8d2d68e11e14e343d6c81b
as that commit was accidentally based on an old patch.

Co-Authored-By: Audrey Dutcher <audrey@rhelmot.io>